Moved models and serivces into their respective directories
