+++ src/app/app.module.ts
@@ -19,7 +19,7 @@ import { OtpVerificationComponent } from './otp-verification/otp-verification.co
 import { EmailService } from './signup/Email.service';
 import { ResetpasswordComponent } from './resetpassword/resetpassword.component';
 import { JobAvailabilityComponent } from './job-availability/job-availability.component';
-import { SignupService } from './signup.service';
+import { SignupService } from './services/signup.service';
 import { HttpClientModule, HttpClient, withFetch } from '@angular/common/http';
 import { provideHttpClient } from '@angular/common/http';
 import { CollegeComponent } from './college/college.component';
@@ -66,7 +66,10 @@ import { AppointmentSchComponent } from './appointment-sch/appointment-sch.compo
     ReactiveFormsModule,
     HttpClientModule,
   ],
-  providers: [EmailService, SignupService, provideHttpClient(), 
+  providers: [
+    EmailService,
+    SignupService,
+    provideHttpClient(),
     // { provide: HttpClient, useFactory: () => provideHttpClient(withFetch()) }
   ],
   bootstrap: [AppComponent],
