+++ src/app/services/student.service.ts
@@ -1,16 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Student } from './student.model'; 
-
+import { Student } from '../models/student.model';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class StudentService {
-
-  private apiUrl = 'http://localhost:8080/student'; 
-  constructor(private http: HttpClient) { }
+  private apiUrl = 'http://localhost:8080/student';
+  constructor(private http: HttpClient) {}
 
   addStudent(student: Student): Observable<Student> {
     return this.http.post<Student>(this.apiUrl, student);
