+++ src/app/student/student.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { StudentService } from '../student.service';
-import { Student } from '../student.model';
+import { StudentService } from '../services/student.service';
+import { Student } from '../models/student.model';
 
 @Component({
   selector: 'app-student',
@@ -40,12 +40,12 @@ export class StudentComponent implements OnInit {
         studentGender: this.studentForm.value.student_gender,
       };
       this.studentService.addStudent(studentData).subscribe(
-        response => {
+        (response) => {
           console.log('Student added:', response);
           alert('Successfully added Student details');
           this.router.navigate(['/college']);
         },
-        error => {
+        (error) => {
           console.error('Error adding student:', error);
         }
       );
